Use unsigned compare for PLIT version-1 check in ReadOtmObj
ReadOtmObj 77.09 -> 77.46.

+++ src/mapobj.cpp
@@ -570,7 +570,7 @@ int CMapObj::ReadOtmObj(CChunkFile& chunkFile)
                     }
                 }
                 chunkFile.PopChunk();
-            } else if (static_cast<int>(chunk.m_version) == 1) {
+            } else if (chunk.m_version == 1) {
                 pointLight->m_color.r = chunkFile.Get1();
                 pointLight->m_color.g = chunkFile.Get1();
                 pointLight->m_color.b = chunkFile.Get1();
